# travel-api/src/main/resources/openapi/travelapp-api.yaml
openapi: 3.1.0
info:
  title: TravelApp API
  version: 1.0.0
  description: |
    API REST de TravelApp. Fuente de verdad — los controllers y DTOs
    de Spring Boot se generan automáticamente desde este fichero
    con openapi-generator en el step `compile` de Maven.
  contact:
    name: TravelApp Team
servers:
  - url: http://localhost:8080/api
    description: Local
  - url: https://api.travelapp.com/api
    description: Producción
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────
# TAGS
# ─────────────────────────────────────────────────────────────
...
paths:
...
  /v1/trips/{tripId}/wishlist:
    parameters:
      - { $ref: '#/components/parameters/TripId' }
    get:
      tags: [Wishlist]
      operationId: listWishlist
      summary: Lista de deseos del viaje
      parameters:
        - name: city
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WishlistItemResponse' }
    post:
      tags: [Wishlist]
      operationId: addWishlistItem
      summary: Añadir ítem a la wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWishlistItemRequest' }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WishlistItemResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── PARAMETERS ─────────────────────────────────────────────
  parameters:
    TripId:
      name: tripId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    WishlistItemResponse:
      type: object
      properties:
        id:              { type: string, format: uuid }
        name:            { type: string }
        category:        { type: string }
        destinationCity: { type: string }
        latitude:        { type: number }
        longitude:       { type: number }
        priority:        { type: integer, enum: [1, 2, 3] }
        rating:          { type: number }
        estimatedCost:   { type: number }
        websiteUrl:      { type: string }
        convertedToEventId: { type: string, format: uuid }
    CreateWishlistItemRequest:
      type: object
      required: [name]
      properties:
        name:            { type: string, minLength: 1 }
        category:        { type: string }
        destinationCity: { type: string }
        latitude:        { type: number }
        longitude:       { type: number }
        priority:        { type: integer, enum: [1, 2, 3], default: 2 }
        externalPlaceId: { type: string }
        websiteUrl:      { type: string }
        estimatedCost:   { type: number }

    # ── Suggestions ───────────────────────────────────────────
...